+++ CHANGES.md
@@ -35,6 +35,9 @@ Unreleased
 * [#229](https://github.com/serokell/xrefcheck/pull/229)
   + Now we call references to anchors in current file (e.g. `[a](#b)`) as
   `file-local` references instead of calling them `current file` (which was ambiguous).
+* [#233](https://github.com/serokell/xrefcheck/pull/233)
+  + Now xrefxcheck does not follow redirect links by default. It fails for permanent
+    redirect responses (i.e. 301 and 308) and passes for temporary ones (i.e. 302, 303, 307).
 
 0.2.2
 ==========